+++ applications/main/subghz/helpers/subghz_custom_event.h
@@ -74,6 +74,7 @@ typedef enum {
     SetTypeKingGatesStylo4k,
     SetTypeBenincaARC,
     SetTypeJarolift,
+    SetTypeDitecGOL4,
     SetTypeANMotorsAT4,
     SetTypeAlutechAT4N,
     SetTypePhoenix_V2_433,