+++ applications/main/subghz/scenes/subghz_scene_set_button.c
@@ -52,6 +52,10 @@ void subghz_scene_set_button_on_enter(void* context) {
         byte_ptr = &subghz->gen_info->jarolift.btn;
         byte_count = sizeof(subghz->gen_info->jarolift.btn);
         break;
+    case GenDitecGOL4:
+        byte_ptr = &subghz->gen_info->ditec_gol4.btn;
+        byte_count = sizeof(subghz->gen_info->ditec_gol4.btn);
+        break;
     case GenNiceFlorS:
         byte_ptr = &subghz->gen_info->nice_flor_s.btn;
         byte_count = sizeof(subghz->gen_info->nice_flor_s.btn);
@@ -101,6 +105,7 @@ bool subghz_scene_set_button_on_event(void* context, SceneManagerEvent event) {
             case GenKingGatesStylo4k:
             case GenBenincaARC:
             case GenJarolift:
+            case GenDitecGOL4:
             case GenNiceFlorS:
             case GenSomfyKeytis:
             case GenSecPlus2: