+++ applications/main/subghz/scenes/subghz_scene_set_counter.c
@@ -58,6 +58,10 @@ void subghz_scene_set_counter_on_enter(void* context) {
         byte_ptr = (uint8_t*)&subghz->gen_info->jarolift.cnt;
         byte_count = sizeof(subghz->gen_info->jarolift.cnt);
         break;
+    case GenDitecGOL4:
+        byte_ptr = (uint8_t*)&subghz->gen_info->ditec_gol4.cnt;
+        byte_count = sizeof(subghz->gen_info->ditec_gol4.cnt);
+        break;
     case GenNiceFlorS:
         byte_ptr = (uint8_t*)&subghz->gen_info->nice_flor_s.cnt;
         byte_count = sizeof(subghz->gen_info->nice_flor_s.cnt);
@@ -141,6 +145,9 @@ bool subghz_scene_set_counter_on_event(void* context, SceneManagerEvent event) {
                 break;
             case GenJarolift:
                 subghz->gen_info->jarolift.cnt = __bswap16(subghz->gen_info->jarolift.cnt);
+                break;
+            case GenDitecGOL4:
+                subghz->gen_info->ditec_gol4.cnt = __bswap16(subghz->gen_info->ditec_gol4.cnt);
                 break;
             case GenNiceFlorS:
                 subghz->gen_info->nice_flor_s.cnt = __bswap16(subghz->gen_info->nice_flor_s.cnt);
@@ -235,6 +242,15 @@ bool subghz_scene_set_counter_on_event(void* context, SceneManagerEvent event) {
                     subghz->gen_info->jarolift.serial,
                     subghz->gen_info->jarolift.btn,
                     subghz->gen_info->jarolift.cnt);
+                break;
+            case GenDitecGOL4:
+                generated_protocol = subghz_txrx_gen_ditec_gol4_protocol(
+                    subghz->txrx,
+                    subghz->gen_info->mod,
+                    subghz->gen_info->freq,
+                    subghz->gen_info->ditec_gol4.serial,
+                    subghz->gen_info->ditec_gol4.btn,
+                    subghz->gen_info->ditec_gol4.cnt);
                 break;
             case GenNiceFlorS:
                 generated_protocol = subghz_txrx_gen_nice_flor_s_protocol(