+++ applications/main/subghz/scenes/subghz_scene_set_serial.c
@@ -58,6 +58,10 @@ void subghz_scene_set_serial_on_enter(void* context) {
         byte_ptr = (uint8_t*)&subghz->gen_info->jarolift.serial;
         byte_count = sizeof(subghz->gen_info->jarolift.serial);
         break;
+    case GenDitecGOL4:
+        byte_ptr = (uint8_t*)&subghz->gen_info->ditec_gol4.serial;
+        byte_count = sizeof(subghz->gen_info->ditec_gol4.serial);
+        break;
     case GenNiceFlorS:
         byte_ptr = (uint8_t*)&subghz->gen_info->nice_flor_s.serial;
         byte_count = sizeof(subghz->gen_info->nice_flor_s.serial);
@@ -137,6 +141,10 @@ bool subghz_scene_set_serial_on_event(void* context, SceneManagerEvent event) {
             case GenJarolift:
                 subghz->gen_info->jarolift.serial = __bswap32(subghz->gen_info->jarolift.serial);
                 break;
+            case GenDitecGOL4:
+                subghz->gen_info->ditec_gol4.serial =
+                    __bswap32(subghz->gen_info->ditec_gol4.serial);
+                break;
             case GenBenincaARC:
                 subghz->gen_info->beninca_arc.serial =
                     __bswap32(subghz->gen_info->beninca_arc.serial);
@@ -171,6 +179,7 @@ bool subghz_scene_set_serial_on_event(void* context, SceneManagerEvent event) {
             case GenKingGatesStylo4k:
             case GenBenincaARC:
             case GenJarolift:
+            case GenDitecGOL4:
             case GenNiceFlorS:
             case GenSecPlus2:
                 scene_manager_next_scene(subghz->scene_manager, SubGhzSceneSetButton);