+++ lib/subghz/protocols/protocol_items.h
@@ -56,3 +56,4 @@
 #include "treadmill37.h"
 #include "beninca_arc.h"
 #include "jarolift.h"
+#include "ditec_gol4.h"