+++ lib/subghz/protocols/public_api.h
@@ -249,6 +249,24 @@ bool subghz_protocol_jarolift_create_data(
     uint16_t cnt,
     SubGhzRadioPreset* preset);
 
+/**
+ * Key generation from simple data.
+ * @param context Pointer to a SubGhzProtocolEncoderDitecGOL4 instance
+ * @param flipper_format Pointer to a FlipperFormat instance
+ * @param serial Serial number
+ * @param btn Button number, 8 bit
+ * @param cnt Counter value, 16 bit
+ * @param preset Modulation, SubGhzRadioPreset
+ * @return true On success
+ */
+bool subghz_protocol_ditec_gol4_create_data(
+    void* context,
+    FlipperFormat* flipper_format,
+    uint32_t serial,
+    uint8_t btn,
+    uint16_t cnt,
+    SubGhzRadioPreset* preset);
+
 typedef struct SubGhzProtocolDecoderBinRAW SubGhzProtocolDecoderBinRAW;
 
 void subghz_protocol_decoder_bin_raw_data_input_rssi(