+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Main changes
 - Current API: 87.6
+* SubGHz: Add **Ditec GOL4** protocol (with programming mode, button switch, add manually) (by @xMasterX (MMX) & @zero-mega)
 * SubGHz: **BFT Mitto fix decode bug** (seed was not resetting after one successful decode)
 * SubGHz: **Somfy Keytis** button switch and **Add Manually support**
 * SubGHz: **KeeLoq** change delta size
 * SubGHz: **Genius Echo/Bravo** add 2 buttons hold simulation (0xB btn code)
 * SubGHz: Signal **Settings Improvements** (PR #968 | by @Dmitry422)
+* SubGHz: KeeLoq **fix display** of **AN-Motors** and **HCS101** keys
 * OFW PR 4338: HID: Fix USB HID keyboard LED state reporting (by @Caballosanex)
 * Apps: Build tag (**22feb2026**) - **Check out more Apps updates and fixes by following** [this link](https://github.com/xMasterX/all-the-plugins/commits/dev)
 ## Other changes